Introduce variable `applicationTeamMap`
This variable maps all instances in the cluster to their responsible
team.

The logic raises an error for any instances that aren't assigned to any
team (after taking into account the `syn.owner` fallback). This error
originates from `teamApplicationMap` which is used in the implementation
to render the contents for `applicationTeamMap`.

Additionally, the logic for `applicationTeamMap` also raises an error if
it detects that any instance is assigned to multiple teams.