Introduce variable `teamApplicationMap`
This variable maps team names to lists of instances. The map only
contains teams which are reponsible for at least one instance in the
cluster.

The logic raises an error if it detects any instances in the cluster
which aren't assigned to any team (after taking into account the
fallback assignment to `syn.owner`).

Usually, this can only happen if the owner team explicitly specifies
that they're not responsible for an application and no other team has
that application in their list.