Extend `allInstances` with the component names of instantiated components
The design doc (SDD-0030) specifies that entries in `syn.<team>.instances`
must be component instances.

However, Commodore doesn't require that instance-aware components generate
an ArgoCD application per instance, and some existing components generate a
single ArgoCD application for all their instances.

For such components, individual instances can't be assigned to a team, but
the whole component can be assigned.

To accommodate this design choice, we keep both the instance and
component names in `allInstances` for instantiated components. This
ensures that library consumers can correctly look up the team for a
component which generates a single ArgoCD application for all instances.